requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-multipart
pillow
selenium
beautifulsoup4
PyQt6
pyinstaller
